requirements.txt
flask
pandas
yfinance
python-dotenv
openai
